Renamed connections2 to connections
