+++ aiospamc/connections.py
